gcc warnings removed

+++ src/position.cpp
@@ -9,8 +9,11 @@ uint64_t zobrist::zobrist_table[NPIECES][NSQUARES];
 //Initializes the zobrist table with random 64-bit numbers
 void zobrist::initialise_zobrist_keys() {
 	PRNG rng(70026072);
-	for (int i = 0; i < NPIECES; i++)
-		for (int j = 0; j < NSQUARES; j++)
+	//gk comparison of integer expressions of different signedness
+	//gk for (int i = 0; i < NPIECES; i++)
+	//gk    for (int j = 0; j < NSQUARES; j++)
+	for (size_t i = 0; i < NPIECES; i++)
+		for (size_t j = 0; j < NSQUARES; j++)
 			zobrist::zobrist_table[i][j] = rng.rand<uint64_t>();
 }
